Move cvss and severity displays into subdirs
Also extract their transform functions into own js modules. This will
allow to resuse them more easily in extra table displayes in future.
