+++ ng/CMakeLists.txt
@@ -208,12 +208,14 @@ set (NG_JS_SRC_FILES
      ${NG_SRC_DIR}/src/web/components/dashboard/legacy/chart/topology.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/controls.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/dashboard.js
-     ${NG_SRC_DIR}/src/web/components/dashboard2/display/cvssdisplay.js
+     ${NG_SRC_DIR}/src/web/components/dashboard2/display/cvss/cvssdisplay.js
+     ${NG_SRC_DIR}/src/web/components/dashboard2/display/cvss/cvsstransform.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/display/datadisplay.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/display/datatable.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/display/display.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/display/displaymenu.js
-     ${NG_SRC_DIR}/src/web/components/dashboard2/display/severityclassdisplay.js
+     ${NG_SRC_DIR}/src/web/components/dashboard2/display/severity/severityclassdisplay.js
+     ${NG_SRC_DIR}/src/web/components/dashboard2/display/severity/severityclasstransform.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/display/utils.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/data/actions.js
      ${NG_SRC_DIR}/src/web/components/dashboard2/data/loader.js
